# src/main/java/dev/ilkersahin/java/spring/gym/dto/auth/LoginRequest.java
package dev.ilkersahin.java.spring.gym.dto.auth;

import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank(message = "Username is required")
        String username,
        @NotBlank(message = "Password is required")
        String password
) {
}