# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Gym CRM API
  description: |
    REST API for Gym Customer Relationship Management system.
    
    ## Authentication
    
    This API uses **JWT (JSON Web Token)** Bearer authentication for protected endpoints.
    
    ### How to authenticate:
    1. Register as a **Trainee** (`POST /api/v1/trainees`) or **Trainer** (`POST /api/v1/trainers`)
    2. Login with your credentials (`POST /api/v1/auth/login`)
    3. Copy the `token` from the response
    4. Click the **Authorize** button (🔓) above
    5. Enter: `Bearer <your-token>` (or just the token - Swagger UI adds "Bearer" prefix)
    6. Click **Authorize** to save
    
    ### Public endpoints (no authentication required):
    - `POST /api/v1/auth/login` - User login
    - `POST /api/v1/trainees` - Register trainee
    - `POST /api/v1/trainers` - Register trainer  
    - `GET /api/v1/training-types` - List training types
    - `GET /health` - Health check
    
    ### Token details:
    - **Format**: JWT (JSON Web Token)
    - **Expiration**: 24 hours
    - **Claims**: `sub` (username), `role` (TRAINEE/TRAINER), `iat`, `exp`
  version: 1.0.0
  contact:
    name: Ilker Sahin
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: |
        Authenticate user and receive JWT token.
        
        **Public endpoint - no authentication required.**
        
        The returned token should be used in the `Authorization` header for protected endpoints:
        ```
        Authorization: Bearer <token>
        ```
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              trainee:
                summary: Trainee login
                value:
                  username: John.Doe
                  password: mySecretPass123
              trainer:
                summary: Trainer login
                value:
                  username: Jane.Smith
                  password: trainerPass456
      responses:
        '200':
          description: Login successful - JWT token returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJKb2huLkRvZSIsInJvbGUiOiJUUkFJTkVFIiwiaWF0IjoxNzA1MzEyMjAwLCJleHAiOjE3MDUzOTg2MDB9.xxx
                username: John.Doe
                role: TRAINEE
        '401':
          description: Invalid credentials - username or password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-01-15T10:30:00"
                status: 401
                error: Unauthorized
                message: Invalid username or password
                path: /api/v1/auth/login
        '429':
          description: Too many failed login attempts - account temporarily locked
          headers:
            Retry-After:
              description: Seconds until the account is unlocked
              schema:
                type: integer
              example: 300
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-01-15T10:30:00"
                status: 429
                error: Too Many Requests
                message: Account temporarily locked due to multiple failed login attempts. Retry after 300 seconds.
                path: /api/v1/auth/login
  /api/v1/auth/logout:
    post:
      tags: [ Authentication ]
      summary: User logout
      description: |
        Invalidate the current JWT token.

        **🔒 Requires authentication**

        After logout, the token cannot be used for any further requests.
        The user must login again to get a new token.

        **Note:** The token is blacklisted until its natural expiration time.
      requestBody:
        description: No request body required - token is extracted from Authorization header
        required: false
        content: { }
      responses:
        '200':
          description: Logout successful - token invalidated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
              example:
                message: Successfully logged out
                username: John.Doe
        '401':
          description: Not authenticated or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2024-01-15T10:30:00"
                status: 401
                error: Unauthorized
                message: Authentication required. Please provide a valid JWT token.
                path: /api/v1/auth/logout

  # ==========================================================================
  # TRAINEES
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.
        
        **How to obtain a token:**
        1. Login via `POST /api/v1/auth/login`
        2. Copy the `token` from the response
        3. Click "Authorize" and enter the token
        
        **Token format:** `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
        
        **Token expiration:** 24 hours
        
        **Token claims:**
        - `sub`: Username
        - `role`: User role (TRAINEE or TRAINER)
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: User's username
          example: John.Doe
        password:
          type: string
          format: password
          description: User's password
          example: mySecretPass123
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token - use in Authorization header as "Bearer <token>"
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJKb2huLkRvZSIsInJvbGUiOiJUUkFJTkVFIn0.xxx
        username:
          type: string
          description: Authenticated user's username
          example: John.Doe
        role:
          type: string
          enum: [TRAINEE, TRAINER]
          description: User's role in the system
          example: TRAINEE
    LogoutResponse:
      type: object
      description: Response after successful logout
      properties:
        message:
          type: string
          description: Confirmation message
          example: Successfully logged out
        username:
          type: string
          description: Username of the logged out user
          example: John.Doe
    # --------------------------------------------------------------------------
    # Trainee
    # --------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Standard error response format
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2024-01-15T10:30:00"
        status:
          type: integer
          description: HTTP status code
          example: 401
        error:
          type: string
          description: HTTP status text
          example: Unauthorized
        message:
          type: string
          description: Detailed error message
          example: Invalid username or password
        path:
          type: string
          description: Request path that caused the error
          example: /api/v1/auth/login